Rename head module
