+++ src/netspresso_trainer/models/registry.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 from .backbones import efficientformer, mobilenetv3_small, mobilevit, resnet50, segformer, vit
 from .full import pidnet
 from .heads.classification import fc
-from .heads.detection import efficientformer_detection_head
+from .heads.detection import faster_rcnn
 from .heads.segmentation import all_mlp_decoder
 
 MODEL_BACKBONE_DICT: Dict[str, Callable[..., nn.Module]] = {
@@ -26,7 +26,7 @@ MODEL_HEAD_DICT: Dict[str, Callable[..., nn.Module]] = {
         'all_mlp_decoder': all_mlp_decoder,
     },
     'detection': {
-        'efficientformer_detection_head': efficientformer_detection_head
+        'faster_rcnn': faster_rcnn
     },
 }
